+++ src/__tests__/modules/workspace.test.ts
@@ -45,14 +45,6 @@ describe('workspace properties', () => {
     expect(cwd).toBe(process.cwd())
   })
 
-  it('should add workspaceFolder', async () => {
-    await helper.edit()
-    let { workspaceFolders, workspaceFolder } = workspace
-    expect(workspaceFolders.length).toBe(1)
-    expect(workspaceFolders[0].name).toBe('coc.nvim')
-    expect(workspaceFolder.name).toBe('coc.nvim')
-  })
-
   it('should check isVim and isNvim', async () => {
     let { isVim, isNvim } = workspace
     expect(isVim).toBe(false)