+++ src/configuration/shape.ts
@@ -3,7 +3,6 @@ import fs from 'fs'
 import { applyEdits, modify } from 'jsonc-parser'
 import path from 'path'
 import { FormattingOptions } from 'vscode-languageserver-types'
-import { URI } from 'vscode-uri'
 import { ConfigurationShape, ConfigurationTarget, IWorkspace } from '../types'
 import { CONFIG_FILE_NAME } from '../util'
 const logger = require('../util/logger')('configuration-shape')
@@ -13,22 +12,16 @@ export default class ConfigurationProxy implements ConfigurationShape {
   constructor(private workspace: IWorkspace) {
   }
 
-  private get nvim(): Neovim {
-    return this.workspace.nvim
-  }
-
   private async modifyConfiguration(target: ConfigurationTarget, key: string, value?: any): Promise<void> {
-    let { nvim, workspace } = this
+    let { workspace } = this
     let file = workspace.getConfigFile(target)
     if (!file) return
     let formattingOptions: FormattingOptions = { tabSize: 2, insertSpaces: true }
-    let content = fs.readFileSync(file, { encoding: 'utf8', flag: 'a+'})
+    let content = fs.readFileSync(file, { encoding: 'utf8', flag: 'a+' })
     value = value == null ? undefined : value
     let edits = modify(content, [key], value, { formattingOptions })
     content = applyEdits(content, edits)
     fs.writeFileSync(file, content, 'utf8')
-    let doc = workspace.getDocument(URI.file(file).toString())
-    if (doc) nvim.command('checktime', true)
     return
   }