+++ src/extensions.ts
@@ -948,7 +948,7 @@ export class Extensions {
       }
       if (rootPatterns && rootPatterns.length) {
         for (let item of rootPatterns) {
-          workspace.addRootPattern(item.filetype, item.patterns)
+          workspace.workspaceFolderControl.addRootPattern(item.filetype, item.patterns)
         }
       }
       if (commands && commands.length) {