+++ src/list/source/folders.ts
@@ -23,11 +23,11 @@ export default class FoldList extends BasicList {
         window.showMessage(`invalid path: ${newPath}`, 'error')
         return
       }
-      workspace.renameWorkspaceFolder(item.label, newPath)
+      workspace.workspaceFolderControl.renameWorkspaceFolder(item.label, newPath)
     })
 
     this.addAction('delete', async item => {
-      workspace.removeWorkspaceFolder(item.label)
+      workspace.workspaceFolderControl.removeWorkspaceFolder(item.label)
     }, { reload: true, persist: true })
 
     this.addAction('newfile', async item => {